+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
 ## Unreleased
-- /
+
+- fixed `minecraftVersion` not being overridden by recipe viewer configs
 
 ## [1.0.0] - 2024-09-01